+++ lib/evision_structurise.ex
@@ -7,12 +7,17 @@ defmodule Evision.Internal.Structurise do
     Evision.Mat.__make_struct__(mat)
   end
 
+  def to_struct(cap = %{:class => :VideoCapture}) do
+    Evision.VideoCapture.__make_struct__(cap)
+  end
+
   def to_struct(tuple) when is_tuple(tuple) do
     Enum.map(Tuple.to_list(tuple), fn elem ->
       to_struct(elem)
     end)
     |> List.to_tuple()
   end
+
   def to_struct(list) when is_list(list) do
     Enum.map(list, fn elem ->
       to_struct(elem)
@@ -21,7 +26,11 @@ defmodule Evision.Internal.Structurise do
   def to_struct(pass_through), do: pass_through
 
   def to_struct_ok(mat = %{:class => :Mat}) do
-    {:ok, Evision.Mat.__make_struct__(mat)}
+    {:ok, to_struct(mat)}
+  end
+
+  def to_struct_ok(cap = %{:class => :VideoCapture}) do
+    {:ok, to_struct(cap)}
   end
 
   def to_struct_ok(tuple) when is_tuple(tuple) do
@@ -35,6 +44,10 @@ defmodule Evision.Internal.Structurise do
   def to_struct_ok(pass_through), do: {:ok, pass_through}
 
   def from_struct(%Evision.Mat{ref: ref}) do
+    ref
+  end
+
+  def from_struct(%{ref: ref}) do
     ref
   end